Handle shims dir that's a child of the current dir - closes #78
When using Vagrant to test cookbooks, it's handy to follow Vagrant
convention and put cookbooks in $PWD/cookbooks. However,
FileUtils#ln_r won't link a directory as a child of itself. Instead of
greatly expanding ln_r to deal with this, instead this patch notices
the situation, swaps the path to a tmpdir, and then moves it back into
place at the end.