Add 'Chef' directive
Given the timeline to eventually remove Chef as a dependency in Berkshelf, this PR introduces the 'chef' directive in the Berksfile, which checks for a specified Chef version on the local machine. You could argue that bundle should take care of this, but I think it makes sense to have this requirement in Berkshelf as well.