Make #info operate off a Berksfile
The original discussion around #info was to look in the CookbookStore. However, there are cookbooks that won't exist in the cookbook store that I want to display info on - specifically when working in a local cookbook where I've either vendored dependencies or using metadata.rb, or any path location for that matter.