add 'chef_servers' key to config and remove 'chef' key
a deprecation will be shown if the chef key is attempted to be assigned. This
will happen when a user loads an old format config file.

the Config#chef function will delegate to the default chef server entry. This
can be removed at a later date and is present just for compatability purposes.