Added a test showing that chef_api_location leaks threads
Added a test showing that chef_api_location leaks threads.  This "thread
leak" becomes apparent very easily on a Windows system when a berks
install downloads 20 to 25 (or so) cookbooks from a chef server (by
which time close to 1000 threads exist in the process).  This problem
also exists on linux, but generally a process on linux can create many
more threads before a problem shows up.  This test currently fails and
will be fixed by code changes to chef_api_location.rb in a subsequent
commit.