Ignore cookbooks from universe if already present
In case a custom location (SCM, path) is specified for a cookbook,
ignore the cookbook version from remote locations like Opscode or a Chef
server.

This is necessary in case a local (potentially development) version uses
different dependencies. The "missing dependency" specified in the
feature is only one possible error scenario, the graph solution can fail
also in other ways.