Only list group-specific dependencies
This commit changes the airty of a number of methods that previously
accepted a list of :only, and :except options. These options are now
honored when the Berksfile object is created, rather than on each method
call.

This is a required change in order to preserve the correct dependency
groups in the lockfile between runs. Otherwise, the Berksfile has no
context of the "true" list of dependencies it needs to install and
manage.