Added some logic so it can handle multiple Github configurations
With this added logic you can use multiple Github organizations and/or
Github Enterprise organizations. This code doesn’t change the default
behaviour, but it does change the github config section from being a
string with a single entry (access_token), to an array with one (just
the one ‘access_token’ entry) or several entries (depending on your
config/setup).