Add the ability to ignore top-level dependencies when removing them
There is a use case, mostly when reducing a lockfile, that you want to
safely remove a nested dependency from a metadata or path location. This
commit adds a new :ignore key to dependency? and remove methods that
tell Berkshelf to ignore the given dependency when checking for deeply
nested dependencies.