Test against STDOUT instead of all output
Recently the safe_yaml gem added a notice printed to STDERR that is
causing this set of tests to fail. This is primarily and issue on Travis
where the version of ruby installed was compiled with an old version of
libyaml but could impact other developers who haven't update libyaml
yet.