Update Vagrantfile generator default vm box
It seems that Chef is now using the bento project to generate their base
boxes for vagrant. These boxes are all listed under
bento/distro-version, instead of chef/distro-version. The current config
for vm.box is causing vagrant up to fail. This commit updates the
config.vm.box variable and an associated test to reflect this change.