Additional generator removal cleanup
The generators got removed already, but the generator_files were never
removed from the repo.  The configs for the generators were also never
deprecated or removed.  Deprecated now so that we can remove them
in the future (with strict config checking on this becomes a hard
error instead of just ignored if we pull the options).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>