[SHACK-295] Kitchen generator conflicts on 'chefignore'
When the Kitchen generator runs it tries to create a 'chefignore' file
that is different from what Berkshelf creates. Because Berkshelf invokes
the kitchen generator by default (if the libraries are present) this
causes the 'chef verify berkshelf --unit' command to fail.

It ends up hanging indefinitely because the 'chefignore' that TK and
Berkshelf create are different, so it prompts the user to resolve this
conflict. This stdin prompt in the automated pipeline causes the tests
to hang forever until they are killed.

This is a partial revert of https://github.com/berkshelf/berkshelf/pull/1702

Signed-off-by: tyler-ball <tball@chef.io>