Allow relative urls in location_path
From original PR:

Hi there!

I'm working on a chef plugin for Nexus Repository, and I ran into some fun!

When using berkshelf to install, it initially hits /universe to get the chef dependency graph. This graph includes a bunch of absolute URLs, which makes for fun times when proxying a repository. I wrote a utility that sets all those URLs to relative, which would in theory work with a tool, and it did for knife (I don't think it uses universe), but berkshelf was not so happy. The reason we prefer relative URLs for Nexus Repo is because once we get into building a group type functionality (grouping multiple repositories under one accessible URL), it makes life a lot easier.

What I did in this PR was check to see if the location_path was set to relative, and if it is, I take the source.uri_string and append the two together, as the source.uri_string SHOULD be the location for Nexus Repo, etc... if your berksfile is configured correctly.

I am probably opening a can of worms with this PR, so feel free to let me know if this is a poorly thought out change.

A few tests appear to be failing but I believe they are unrelated to this change.

Signed-off-by: Jeffry Hesse <jeffryxtron@gmail.com>