Remove the use of Dir.chdir in the code
The use of Dir.chdir is bad because it is global state and is not
threadsafe.  In ruby-3.0 Dir.chdir use from another thread from
within a Dir.chdir is now a hard error.  It was unclear exactly
which Dir.chdir was problematic, so they were all removed.

This may also clean up some issues with test-kitchen where Berkshelf
may be run in threads and not be threadsafe.

(Test-kitchen should still fork off berkshelf and stop using it in
threads)

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>